Forgot to add the 3dStudio MAX icons
These are used in some downstream CAD code and I forgot to add them to the toolkit for wider access. As this is by far the best icon for 3D Studio files, at least in terms of holding up at the smallest 16 pixel size, it is helpful to have it generally available in this toolkit.